# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Hotels API's",
        "contact": {
            "name": "Rami Badran",
            "email": "ramibadran.82@gmail.com"
        },
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "schemes": [
        "http"
    ],
    "paths": {
        "/token": {
            "post": {
                "tags": [
                    "Token"
                ],
                "summary": "Generate API Token",
                "operationId": "token",
                "parameters": [
                    {
                        "name": "username",
                        "in": "formData",
                        "description": "client username",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "client password",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "USER-IP",
                        "in": "header",
                        "description": "User ip",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "USER-DEVICE-AGENT",
                        "in": "header",
                        "description": "Device agent . EX : IOS,Android",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "success response with custom message"
                    },
                    "200": {
                        "description": "success response and token not created with custom message"
                    },
                    "401": {
                        "description": "unauthorized operation with custom message"
                    },
                    "412": {
                        "description": "precondition failed with custom message"
                    },
                    "503": {
                        "description": "service unavailable with custom message"
                    }
                }
            }
        },
        "/hotels": {
            "get": {
                "tags": [
                    "Hotels"
                ],
                "summary": " hotels from third-part providers based on users filter queries",
                "operationId": "getHotels",
                "parameters": [
                    {
                        "name": "city",
                        "in": "query",
                        "description": "IATA code (AMM),use my sample AMM",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "fromDate",
                        "in": "query",
                        "description": "date from,format ISO_LOCAL_DATE,use my sample 2019-2-12",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "toDate",
                        "in": "query",
                        "description": "date to,format ISO_LOCAL_DATE,use my sample 2019-2-15",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "numberOfAdults",
                        "in": "query",
                        "description": "number of adults,use my sample 2",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success response with custom message"
                    },
                    "401": {
                        "description": "unauthorized operation with custom message"
                    },
                    "503": {
                        "description": "service unavailable with custom message"
                    }
                }
            }
        }
    },
    "definitions": {}
}